Improve scfp calculation algorithm; add cube file export; add d-orbitals calculation

+++ src/cpp/Atom.h
@@ -9,7 +9,7 @@ public:
     Atom(const std::array<double, 3>& position,
          const std::string& label = "",
          int number = 0,
-         double charge = 0.0);
+         int charge = 0);
 
     std::string repr() const;
 
@@ -17,13 +17,13 @@ public:
     const std::array<double, 3>& get_position() const { return position_; }
     const std::string& get_label() const { return label_; }
     int get_number() const { return number_; }
-    double get_charge() const { return charge_; }
+    int get_charge() const { return charge_; }
 
 private:
     std::array<double, 3> position_;
     std::string label_;
     int number_;
-    double charge_;
+    int charge_;
 };
 
 #endif // ATOM_H
